requirements.txt
streamlit
langchain
faiss-cpu
sentence-transformers
PyMuPDF
python-docx
huggingface_hub
